Add files via upload
ERC-2981 is an Ethereum token standard that introduces a royalty payment feature to tokens, allowing creators to earn a percentage of the value of each subsequent transaction involving their token. This feature can be used for various purposes, such as incentivizing creators to produce high-quality content, promoting collaboration among creators, and encouraging innovation in the token ecosystem.

The ERC-2981 standard specifies the following functions:

royaltyInfo(uint256 _tokenId, uint256 _salePrice): This function returns the recipient and percentage of the royalty payment that should be sent to the creator of the token with the specified _tokenId and _salePrice. The _salePrice parameter is the price of the token at the time of sale.

setRoyaltyRecipient(address _recipient, uint256 _percentage): This function sets the royalty recipient and percentage for the token. The _recipient parameter is the address of the account that will receive the royalty payment, and the _percentage parameter is the percentage of the value of each subsequent transaction that will be paid as royalty.

royaltyReceived(address _royaltyRecipient, address _buyer, uint256 _tokenId, uint256 _value): This function is called by the ERC-721 token contract when a token is sold, and it sends the royalty payment to the _royaltyRecipient account.

By using the ERC-2981 standard, creators can receive a portion of the value of their tokens even after they have been sold or traded on secondary markets. This incentivizes them to continue creating and improving their content, and it can also help to build a more sustainable token ecosystem.